+++ docs/changelog.rst
@@ -1,6 +1,13 @@
 Changelog
 ~~~~~~~~~
 
+0.22.0 (September 2022)
+-----------------------
+
+* Fix tracking of qubit indices on devices with non-contiguous qubit sets.
+* Update dependencies.
+* Updated pytket version requirement to 1.6.
+
 0.21.0 (July 2022)
 ------------------